containerize demos (second try)
