+++ container/.mocharc.jsonc
@@ -1,0 +1,10 @@
+{
+  // To run tests written in TypeScript without having to compile them, we need to run them in ts-node
+  "require": ["ts-node/register"],
+
+  // Any *.test.ts file in ./test will be run as a test
+  "spec": "test/**/*.test.ts",
+
+  // A change in sources or tests should trigger test re-run
+  "watch-files": ["test/**", "src/**"]
+}
