+++ container/tsconfig.json
@@ -1,10 +1,14 @@
 {
   "compilerOptions": {
-    "target": "es2020",
-    /* Recommend by Webpack. */
-    "module": "es2020",
-    /* Needed with module: es2020 or else stuff breaks. */
-    "moduleResolution": "node",
+    /* Use . as rootDir so build contains a nested src folder.
+     * This keeps things consistent across different projects,
+     * some of which have a "generated" folder (thus causing
+     * Typescript to set rootDir = . by default) while
+     * others don't.
+     */
+    "rootDir": ".",
+    "outDir": "build",
+    "target": "es2019",
     /* Enable strict type checking. */
     "strict": true,
     /* The workspace structure causes tsc to check all
@@ -14,16 +18,15 @@
      * to skipLibChecks.
      */
     "skipLibCheck": true,
+    "module": "commonjs",
     /* Enable interop with dependencies using different module systems. */
     "esModuleInterop": true,
-    /* Include generated .js files in compilation run */
-    "allowJs": true,
-    /* We don't need to emit declarations. */
-    "declaration": false,
+    /* Emit .d.ts files. */
+    "declaration": true,
     /* Emit sourcemap files. */
     "sourceMap": true,
     /* Use tslib to reduce generated js file sizes. */
     "importHelpers": true
   },
-  "include": ["src/site", "generated/site"]
+  "include": ["src"]
 }
