+++ demos/src/site/containers/counter.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+
+    <title>Testing out bundle</title>
+  </head>
+
+  <body>
+    <!-- HTML page title -->
+    ADD-ONLY COUNTER DEMO
+
+    <!-- HTML page variables and buttons -->
+    <p id="counter">0</p>
+    <button id="increment">💯️</button>
+    <button id="decrement">-💯️</button>
+    <br />
+    <button id="reset">Reset</button>
+    <!-- TODO <button id="strongReset">Strong Reset</button> -->
+
+    <!-- Run main script -->
+    <script src="counter.js"></script>
+
+    <!-- Add more scripts -->
+  </body>
+</html>
