+++ demos/src/site/containers/minesweeper.html
