--- demos/src/site/matrix/plaintext.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-
-    <style>
-      textarea {
-        font-size: 12px;
-      }
-    </style>
-
-    <title>Plain text editor</title>
-  </head>
-
-  <body>
-    <textarea id="textarea" rows="20" cols="50"></textarea>
-
-    <script src="plaintext.js"></script>
-  </body>
-</html>
