--- demos/src/site/matrix/whiteboard.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-
-    <title>Whiteboard demo</title>
-  </head>
-
-  <body>
-    <!-- HTML page title -->
-    Whiteboard Demo
-
-    <!-- HTML page variables and buttons -->
-    <p></p>
-    <div class="btn-colors">
-      <button id="red" style="background-color: red"></button>
-      <button id="orange" style="background-color: orange"></button>
-      <button id="yellow" style="background-color: yellow"></button>
-      <button id="green" style="background-color: green"></button>
-      <button id="blue" style="background-color: blue"></button>
-      <button id="purple" style="background-color: purple"></button>
-      <button id="black" style="background-color: black"></button>
-    </div>
-
-    <div class="btn-clear">
-      <button id="clear">Clear</button>
-    </div>
-
-    <canvas id="board" width="1000" height="750"></canvas>
-
-    <!-- Run main script -->
-    <script src="whiteboard.js"></script>
-  </body>
-
-  <style>
-    .btn-colors button {
-      height: 20px;
-      width: 20px;
-    }
-
-    #board {
-      border: 2px solid black;
-    }
-  </style>
-</html>
