--- demos/src/site/matrix/whiteboard.ts
@@ -1,124 +1,0 @@
-import * as crdts from "compoventuals";
-import { MatrixWidgetNetwork } from "compoventuals-matrix-widget";
-import { min, max, round } from "mathjs";
-import $ from "jquery";
-
-let client = new crdts.Runtime(
-  new MatrixWidgetNetwork("com.herokuapp.compoventuals-tests.whiteboard"),
-  {
-    periodMs: 300,
-  }
-);
-
-// The key represents a point in the form: x:y
-// The value is the color of the stroke.
-let clientBoard: crdts.LwwCMap<string, string> = client.registerCrdt(
-  "whiteboardId",
-  new crdts.LwwCMap()
-);
-
-window.onload = function () {
-  var colors = document.getElementsByClassName("btn-colors");
-  var clear = <HTMLButtonElement>document.getElementById("clear");
-  var board = <HTMLCanvasElement>document.getElementById("board");
-  var ctx = board.getContext("2d");
-  let gran = 2;
-
-  let roundGran = function (n: number): number {
-    return round(n / gran) * gran;
-  };
-
-  let interpolate = function (sX: number, sY: number, eX: number, eY: number) {
-    // special case - line goes straight up/down
-    if (sX == eX) {
-      let pts = [];
-      for (
-        let i = roundGran(min(sY, eY));
-        i <= roundGran(max(sY, eY));
-        i += gran
-      ) {
-        pts.push(roundGran(sX) + ":" + i);
-      }
-
-      return pts;
-    }
-
-    let slope = (eY - sY) / (eX - sX);
-    let pts = [];
-    let intercept = sY - slope * sX;
-
-    // Depending on slope, iterate by xs or ys
-    if (slope <= 1 && slope >= -1) {
-      for (
-        let i = roundGran(min(sX, eX));
-        i <= roundGran(max(sX, eX));
-        i += gran
-      ) {
-        pts.push(i + ":" + roundGran(slope * i + intercept));
-      }
-    } else {
-      for (
-        let i = roundGran(min(sY, eY));
-        i <= roundGran(max(sY, eY));
-        i += gran
-      ) {
-        pts.push(roundGran((i - intercept) / slope) + ":" + i);
-      }
-    }
-
-    return pts;
-  };
-
-  // Draw points
-  clientBoard.on("Set", (event) => {
-    var keys = event.key.split(":");
-    ctx!.fillStyle = clientBoard.get(event.key)!;
-    ctx!.fillRect(parseInt(keys[0]), parseInt(keys[1]), gran, gran);
-  });
-
-  // Clear points
-  clientBoard.on("Delete", (event) => {
-    var keys = event.key.split(":");
-    ctx!.clearRect(parseInt(keys[0]), parseInt(keys[1]), gran, gran);
-  });
-
-  // Mouse Event Handlers
-  if (board) {
-    var ctx = board.getContext("2d");
-    var color = "black";
-
-    var isDown = false;
-    var canvasX: number, canvasY: number, prevX: number, prevY: number;
-
-    // Update color selection
-    $(colors).on("click", function (e: JQuery.ClickEvent) {
-      color = e.target.id;
-    });
-
-    $(clear).on("click", function () {
-      clientBoard.reset();
-    });
-
-    // Draw on board
-    $(board)
-      .on("mousedown", function (e: JQuery.MouseDownEvent) {
-        isDown = true;
-        prevX = e.pageX - board.offsetLeft;
-        prevY = e.pageY - board.offsetTop;
-      })
-      .on("mousemove", function (e: JQuery.MouseMoveEvent) {
-        if (isDown !== false) {
-          canvasX = e.pageX - board.offsetLeft;
-          canvasY = e.pageY - board.offsetTop;
-          interpolate(prevX, prevY, canvasX, canvasY).forEach(function (pt) {
-            clientBoard.set(pt, color);
-          });
-          prevX = canvasX;
-          prevY = canvasY;
-        }
-      })
-      .on("mouseup", function () {
-        isDown = false;
-      });
-  }
-};
