+++ demos/src/site/non_container_demos/counter_webrtc.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+
+    <title>Testing out bundle</title>
+  </head>
+
+  <body>
+    <!-- HTML page title -->
+    ADD-ONLY COUNTER DEMO
+
+    <!-- HTML page variables and buttons -->
+    <p id="counter">0</p>
+    <button id="increment">💯️</button>
+    <button id="decrement">-💯️</button>
+
+    <!-- Run main script -->
+    <script src="counter_webrtc.js"></script>
+
+    <!-- Add more scripts -->
+  </body>
+</html>
