--- demos/src/site/variablecounter.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Title</title>
-  </head>
-  <body>
-    <p id="variable_counter">0</p>
-    <label for="variable_amount"> Amount </label>
-    <br />
-    <input type="number" id="variable_amount" />
-    <button id="variable_increment">Increment</button>
-    <button id="variable_decrement">Decrement</button>
-    <button id="variable_reset">Reset</button>
-    <script src="variablecounter.js"></script>
-  </body>
-</html>
