+++ demos/webpack.config.ts
@@ -6,17 +6,19 @@ const config: webpack.Configuration = {
   mode: "development",
   devtool: "inline-source-map",
   entry: {
-    counter: "./src/site/counter.ts",
-    whiteboard: "./src/site/whiteboard.ts",
-    aspace: "./src/site/aspace.ts",
-    counter_webrtc: "./src/site/counter_webrtc.ts",
-    minesweeper: "./src/site/minesweeper.ts",
-    variablecounter: "./src/site/variablecounter.ts",
-    text: "./src/site/text.ts",
-    plaintext: "./src/site/plaintext.ts",
-    "matrix/whiteboard": "./src/site/matrix/whiteboard.ts",
-    "matrix/minesweeper": "./src/site/matrix/minesweeper.ts",
-    "matrix/plaintext": "./src/site/matrix/plaintext.ts",
+    "containers/aspace": "./src/site/containers/aspace.ts",
+    "containers/counter": "./src/site/containers/counter.ts",
+    "containers/minesweeper": "./src/site/containers/minesweeper.ts",
+    "containers/plaintext": "./src/site/containers/plaintext.ts",
+    "containers/text": "./src/site/containers/text.ts",
+    "containers/whiteboard": "./src/site/containers/whiteboard.ts",
+    "hosts/plain": "./src/site/hosts/plain.ts",
+    "non_container_demos/counter_matrix":
+      "./src/site/non_container_demos/counter_matrix.ts",
+    "non_container_demos/counter_webrtc":
+      "./src/site/non_container_demos/counter_webrtc.ts",
+    "non_container_demos/counter_ws":
+      "./src/site/non_container_demos/counter_ws.ts",
     // TODO: Add more entries as needed for new pages
   },
   output: {
@@ -51,29 +53,12 @@ const config: webpack.Configuration = {
   plugins: [
     new CopyWebpackPlugin({
       patterns: [
-        { from: "./src/site/index.html", to: "index.html" },
-        { from: "./src/site/counter.html", to: "counter.html" },
-        { from: "./src/site/whiteboard.html", to: "whiteboard.html" },
-        { from: "./src/site/counter_webrtc.html", to: "counter_webrtc.html" },
-        { from: "./src/site/minesweeper.html", to: "minesweeper.html" },
-        { from: "./src/site/aspace.html", to: "aspace.html" },
-        { from: "./src/site/variablecounter.html", to: "variablecounter.html" },
-        { from: "./src/site/text.html", to: "text.html" },
-        { from: "./src/site/plaintext.html", to: "plaintext.html" },
+        // Copy all HTML files to the build dir.
         {
-          from: "./src/site/matrix/whiteboard.html",
-          to: "matrix/whiteboard.html",
+          from: "./src/site/",
+          to: "./",
+          filter: (resourcePath: string) => resourcePath.endsWith(".html"),
         },
-        {
-          from: "./src/site/matrix/minesweeper.html",
-          to: "matrix/minesweeper.html",
-        },
-        {
-          from: "./src/site/matrix/plaintext.html",
-          to: "matrix/plaintext.html",
-        },
-        { from: "./src/site/reset.html", to: "reset.html" },
-        // TODO: Add more entries as needed for new pages
       ],
     }),
   ],
